Add files via upload
these are Multi tables Dataset and a collection of structured data tables designed for research and analysis purposes. This dataset contains multiple tables, each serving a specific purpose and collectively providing a comprehensive view of a hypothetical retail business and entertainment. 
this database contains 11 tables which was carefully analysed.
The purpose of this analysis was to categorically view how sales was impacted and at what point it grew to its highest and also the customers interests.